db: move delete-only compaction code into compaction_delete.go
Move code specific to delete-only compactions into compaction_delete.go.
